Integrate Nextcloud 16 and app version overrides
We do the version overrides on the NixOS module level, because we
already have all information about the individual apps available in the
upstream.json files.

The implementation basically merges the sets for version 15 and 16 and
allows to use apps targetted at Nextcloud 15 in version 16, but not vice
versa, because versions older than 16 do not support overriding the
version check.

Since we now have Nextcloud 15 as the oldest version available, we can
now remove the special "old" package, which was only used by the upgrade
test, because we now have a way to perform upgrades between 15, 16 and
soon 17.

In the future, upgrading to a new major version should be simply a
matter of adding the directory to the packages/ directory and run
update.sh.

Users of Nextcloud 15 will get an appropriate error if they build a
system configuration that uses apps which aren't supported and there is
a new option nextcloud.majorVersion, which allows to set the version to
an older one.

So with this, the Nextcloud 16 package that has been added a few changes
ago is now fully integrated.

Signed-off-by: aszlig <aszlig@nix.build>