Add a release.nix for Hydra
So far I ran all the tests manually, which is a bit tedious to do,
especially when it comes to the upgrade test which takes a long time to
complete.

What's missing at the moment is a proper channel, but since I still
consider this repository experimental, I don't see a need to add one.

Signed-off-by: aszlig <aszlig@nix.build>