manual.nix: Fix evaluation in restricted mode
If pkgs.path is used within a string context, the Nix evaluator will add
its contents to the Nix store. However, we don't want this to happen
whenever we only evaluate nixpkgs, because it involve copying the full
tree (again) just in order to evaluate something that's accessible in
non-store paths.

Signed-off-by: aszlig <aszlig@nix.build>