Add NixOS module for OSRM
This is useful for Nextcloud Maps, so that people are able to host their
own routing backend. Unfortunately, running OSRM with a very large
dataset requires a LOT of memory, so I'm not enabling this by default if
the app is enabled.

The implementation patches Nextcloud Maps to use Unix domain sockets to
access the OSRM backend, so that we can properly sandbox the service and
don't allow any outgoing connections.

Right now, I've included a few datasets, namely "planet", "europe" and
"germany" for now, because they're the most useful ones at least in my
deployments. The long-term goal however is to implement updating and
gathering of these datasets into our updater.

One thing however that is always enabled whenever the "maps" app is
enabled, is the proxy-routing patch, which makes sure that routing is
never sent via the user's browser but proxied through Nextcloud Server.

The patch is basically a rebase of the one I submitted upstream at
https://github.com/nextcloud/maps/pull/167 against Maps version 0.1.2.

Signed-off-by: aszlig <aszlig@nix.build>