osrm: Switch osmDataset option to use PBF files
Unfortunately, the upstream mirrors frequently purge old data, so it's
not really feasible to provide this data on our own but let users
directly specify the data file they want to use.

This also frees us of a way to regularily update the data and instead
shift it to be the user's responsibility. Fortunately, having outdated
data here only gives you outdated routing information and shouldn't have
any security impact.

Signed-off-by: aszlig <aszlig@nix.build>