updater: Allow Nextcloud server to be non-existant
When updating Nextcloud to a new major version, all that we need to do
now is just creating a new directory with the major version in packages/
and run the updater.

So far this has been a bit tricky because we did have a fallback to just
use the major version as the current version in the Nextcloud structure,
but if the current Nextcloud version is something like 17.0.0, it won't
update at all.

The "version" attribute is now an Optional instead and whenever it is
None, we just use the major version with a caret for the semver
specification, so we *always* get the latest version for that major
release.

Signed-off-by: aszlig <aszlig@nix.build>