nixos: Use literalExample for defaultText attrs
If we don't use literalExample here, the default values that end up in
the manual will get quotes around the values, which would indicate that
the default is a string.

This however is not the case everywhere we use defaultText, so we're now
using literalExample so that the expression is used verbatim and this no
longer suggests that it might be a string value.

Signed-off-by: aszlig <aszlig@nix.build>