manual: Fix build on NixOS 19.09
On NixOS 19.09 the generation of the options in (among others like JSON)
DocBook format have been factored out into pkgs.nixosOptionsDoc and
subsequently also changed the path of the XSL files.

While there is still a lot of stuff that we duplicate on our side, we
now use the existence of nixosOptionsDoc function to indicate whether
we're on 19.03 or on 19.09 and subsequently use the right path.

I also dropped support code for NixOS 18.03, because we really don't
want to support anything older than 19.03 and even 19.03 is going to be
dropped soon, because this very project is not yet officially
"released".

Signed-off-by: aszlig <aszlig@nix.build>