libreoffice-online: Explicitly use GCC 7 for now
In NixOS 19.09 or newer, LibreOffice Online no longer compiles because
GCC 8 is the new default compiler.

Ideally we should upgrade LibreOffice Online to a new upstream version
and I already have a branch working on that. But until that branch is
ready to merge, let's at least make sure the old version still compiles.

Signed-off-by: aszlig <aszlig@nix.build>