Use exec prefixes instead of PermissionsStartOnly
PermissionsStartOnly is deprecated since systemd 240, but it's still
available for backwards-compatibility.

Now we could have kept using it, but the PermissionsStartOnly option is
now equivalent to the "+" executable prefix, which has the following
description:

  If the executable path is prefixed with "+" then the process is
  executed with full privileges. In this mode privilege restrictions
  configured with User=, Group=, CapabilityBoundingSet= or the various
  file system namespacing options (such as PrivateDevices=, PrivateTmp=)
  are not applied to the invoked command line (but still affect any
  other ExecStart=, ExecStop=, ... lines).

Since we *only* want to affect User and Group in the cases where we use
it but leave everything else alone, we can't simply use
PermissionsStartOnly anymore.

I haven't found out why the behaviour has changed and also which systemd
version was responsible for this, a bisect against nixpkgs however
showed that it has to be either version 240 or version 241.

Nevertheless, since our tests fail since systemd version 241, I'd say
it's safe to infer that the behaviour has actually changed, so let's use
the "!" executable prefix which *only* affects User and Group.

Fortunately, these prefixes are available in systemd 239 as well, so we
still support NixOS 19.03 with this.

Signed-off-by: aszlig <aszlig@nix.build>