Add an option to enable/disable the whole service
Usually you would just enable/disable the service by just either adding
the module to the "imports" attribute or removing it from there.

This would indicate that we don't really need an "enable" option, but
when it comes to tests, where we globally import the main module, we
usually only want to enable the service on *one* node only.

I also modified make-test.nix to enable the service by default if there
is only one node, so we don't need to add redundant "enable = true"
lines in our test configurations.

Signed-off-by: aszlig <aszlig@nix.build>