tests: Add a small test for Talk with coturn
The idea here is that we have four browsers that are joining a common
Nextcloud Talk conversiation and we just wait until the connection is
done and we then eg. OCR the screenshots of all the peers, looking
whether every browser gets the camera output of all the peers.

Camera output is provided by v4l2loopback and we just use a small helper
script that is providing the video frames, which are generated in a way
where we roughly see how much time has passed since the helper script
has started providing frames. The frames are also labelled with a text
in the center, which we can later OCR (hence the use of Inconsolata,
because it's easy to OCR with less errors).

Unfortunately, I haven't managed to get reliable calls even with a TURN
server and I suspect bandwidth issues here. On the other hand even when
using Nextcloud Talk on production machines, I've encountered
reliability issues as well even for 3 peers with broadband connections.

With our test however, the reliability is worse in a similar vein even
though we use virtio and vde_switch, so in theory it should perform
better.

So right now, the test here is not making too many assertions and I also
didn't add the OCR part but instead only check whether we have more than
two video streams via WebDriver.

Since we've essentially fixed the coturn server in the previous commit,
I'm going to check whether this really has improved reliability on my
test deployments and maybe figure out whether either my test setup is
wrong or whether this really is an issue that upstream has as well.

Signed-off-by: aszlig <aszlig@nix.build>