Update all Nextcloud apps
Apps updated:

  ocsms (1.10.1 -> 2.1.6):

    Changes for version 2.1.6:

      * Fix invalid code missed on review

    Changes for version 2.1.5:

      * Fix info.xml validation error

  oidc_login (1.0.0 -> 1.0.1):

    * Regenerate composer autoload files

  passwords (2019.11.0 -> 2019.11.1):

    * Fix local dictionary password generator issues

The update of oidc_login fixes the conflict with the sociallogin app
that I discovered in 2459f067ce782b69df22514c0295df42bf2818b1.

So I subsequently removed the sociallogin app from the excludes list in
our upgrade test.

Thanks to @pulsejet and @zorn-v for the quick reaction.

Signed-off-by: aszlig <aszlig@nix.build>