tests/upgrade: Exclude "cms_pico" app
This one unfortunately needs write access to its own app directory, but
in our setup, all app directories are read-only.

We might want to patch this in the future though to use a different
path, but since I'm not using this app myself, I decided to exclude it
for now from the test.

Signed-off-by: aszlig <aszlig@nix.build>