lool: Patch Poco to not set a default curve
When using services.nginx.recommendedTlsSettings, it will result in the
following entries in the nginx configuration:

  ssl_session_cache shared:SSL:42m;
  ssl_session_timeout 23m;
  ssl_ecdh_curve secp384r1;
  ssl_prefer_server_ciphers on;
  ssl_stapling on;
  ssl_stapling_verify on;

The important part here is the curve selection here, which will default
to the prime256v1 curve and subsequently fail to do the TLS handshake.

Since curve selection is no longer necessary in OpenSSL 1.1, let's make
sure we don't select one here and patch Poco to avoid selection if the
OpenSSL version is recent enough.

Note that we do not need to apply the same patch for loolwsd, since we
don't use TLS there and instead rely on nginx instead.

We also wouldn't need to have any encryption at all, even for the Poco
part, since that part is only responsible for connecting from loolwsd to
the Nextcloud server and we do that via an internal Unix domain socket.

Unfortunately, we can't rid of it since Nextcloud enforces a certain
base URL and patching Nextcloud itself would be a much larger patch than
what we've done here.

Signed-off-by: aszlig <aszlig@nix.build>