updater: Handle invalid certs for new apps
So far we just used the old app specification if we got an update for an
app with an invalid certificate.

However, when updating a few minutes ago, there was a new "mindmap_app"
Nextcloud app, which had a revoked certificate. This led to the updater
to fail because there is no old version of the app.

What we do now is just remove the entry of the old app whenever there is
a new app with an invalid certificate.

Signed-off-by: aszlig <aszlig@nix.build>