updater: Only try to update existing themes
Since the breeze dark theme we had so far is now an ordinary Nextcloud
app, there is no branch anymore for Nextcloud 18, so we no longer need
to fetch it seperately.

To make sure we still get updates for the old major versions of
Nextcloud, we now fetch themes only if they're already defined in
upstream.json, so we can phase it out once we no longer support
Nextcloud 15 and 16.

Signed-off-by: aszlig <aszlig@nix.build>