packages/16: Blacklist mindmap_app version 0.0.1
The app has a revoked certificate, so let's actually avoid fetching it
by constraining the version to "greater than 0.0.1". While I highly
doubt that there ever will be a new version, since the app already
exists as "files_mindmap" and I believe that "mindmap_app" was submitted
in error, let's err on the safe side and only blacklist this specific
version.

Signed-off-by: aszlig <aszlig@nix.build>