Remove XMPP server for ojsxc
The build for MongooseIM has been broken since NixOS 19.09 because of
some changes in the Erlang ecosystem. I currently don't have time to
investigate this and since this is breaking our Hydra builds I decided
to remove it for now.

Besides, the ojsxc Nextcloud app also doesn't declare compatibility with
Nextcloud 17 and 18, so most users probably won't install it.

Also, in the long term I guess it's better to have a bridge between
Nextcloud Talk and XMPP rather than trying to add an app for XMPP that
more or less feels redundant.

Signed-off-by: aszlig <aszlig@nix.build>