packages/17: Fix tarball URL for cospend
Unfortunately, we don't handle unknown file extensions with fetchzip
very well, so we'd need to either explicitly specify the right extension
or file format or use something like file(1) to determine the type.

Fortunately however, it seems that cospend is the only app where we have
a different file name in the URL and since upstream already has changed
the URL in the app store to point to the GitLab URL, we can simply
change the URL here and be done. The hashes are also still the same.

Of course, if we have something like this again in the future, we'd
probably need a way to provide the right extension, but for now, let's
just get rid of this small annoyance.

Signed-off-by: aszlig <aszlig@nix.build>