packages: Add alwaysEnable meta for shipped apps
There is a distinction between a shipped app that is enabled by default
and a shipped app that is enabled by default but can't be disabled.

So far we've ignored this in our updater, but if a user disables one of
these apps, the app enabler/disabler will throw a runtime error and we'd
like to avoid those.

The reason why I didn't use the readOnly attribute on the NixOS options
of the corresponding apps is because we'd run into an eval error even if
we set the enable option to true, because it has already been defined in
the option declaration.

Using assertions instead will only result in an evaluation error if one
of these apps is *really* set to be disabled.

Signed-off-by: aszlig <aszlig@nix.build>