updater: Add assertion for Nextcloud version
When exporting data to the corresponding JSON file, we really want to
ensure that we got a version for the Nextcloud server.

The reason why the version field is nullable in the first place is
because we actually *want* to support a non-existing Nextcloud version
during import, which is useful when adding a new major version.

Signed-off-by: aszlig <aszlig@nix.build>