updater: Show changelogs after updating everything
While we could have just gathered all the changelogs in a big list of
strings and join them before outputting, I decided to not do it that way
and instead re-arrange and refine some of our types.

We now nave a new AppChanges type, which essentially is a condensed
version of the App type and it only contains all the information that is
necessary for determining which apps have been added/removed and which
versions got updated along with the changelogs.

Right now this new type more or less results in the same functionality
as concatenating the list of changes, but the reason why I introduced
this dedicated type is because this is only laying the foundation for a
few more changes where we actually deduplicate all the changelogs.

The reason why we want to deduplicate them is because so far every
update that consists of several app updates is costing me hours of work
to not only figure out the changes but check which apps are part of
which major versions and/or whether the actual changes are the same but
they occur in different versions.

Something like the latter is pretty common among a bunch of apps, which
for example have a version 1.2.1 for Nextcloud 15 and a version 1.3.1
for Nextcloud 16. In this case both would have the same changes but
they're just applied to a different base. In our update commits we want
to have such changes merged into a single paragraph, but doing so would
mean that we need to have a way to easily compare app versions *across*
Nextcloud major versions.

Signed-off-by: aszlig <aszlig@nix.build>