updater: Group changelogs by app and major
Especially when there are a lot of updated apps, the whole changelist
gets quite big with lots of duplicates. This always was a bit tedious to
manually sort them out and thus provide a more relevant changelog in the
update commit.

So we now group all the changes based on the corresponding Nextcloud
major version, so that if for example app A is updated for more than one
Nextcloud major version we actually only get *one* entry for app A.

The implementation still has a few rough edges, but this is already way
less cumbersome than how we handled it before.

Signed-off-by: aszlig <aszlig@nix.build>