Use Redis for file locking and memory caching
While it's not strictly necessary to switch from APCu to Redis for eg.
file paths, it helps to use a consistent caching mechanism.

The main reason for doing this is because with PostgreSQL file locking
is really buggy and there are a multitude of upstream issues around
that. Given that even upstream is dogfeeding on Redis, it's probably a
better idea to switch to that, since it most probably will always be
better supported than PostgreSQL.

Additionally, I'm not using the Redis module from NixOS, because it
doesn't support socket activation and also is not confined. The reason
why I'm using ip2unix for the Redis process is because Redis itself
doesn't support systemd socket activation, so we're listening to TCP and
ip2unix simply picks up the Unix socket accordingly.

Signed-off-by: aszlig <aszlig@nix.build>