module: Set redis.port to 0 for NC <= 16 compat
Since Nextcloud 16 didn't yet detect whether the host name is a Unix
socket (starts with a '/'), the host the port argument passed to the
Redis PHP extension is the default port. This in turn cases the
extension to try connecting via TCP.

When looking at the implementation of the extension, the check for Unix
sockets looks like this:

  if (address[0] == '/' && redis_sock->port < 1) {
    ...
  }

So all we need to do here is to pass port 0 to get a Unix domain socket
connection.

Signed-off-by: aszlig <aszlig@nix.build>