lool: Rebase patch for "richdocuments" app
Unfortunately, version 3.5.3 was tagged incorrectly in the upstream
project, so after rebasing the patch against the "v3.5.3" upstream tag I
was first a bit confused about why the patch didn't apply.

A diff of the tarball against the tagged version also confirmed this and
after reporting this upstream[1], @juliushaertl pointed out that there
is indeed a mismatch:

> Ok, apparently I tagged the master branch instead of the stable-3.5
> one, sorry about the confusion there. I'll see if I can update the tag
> later.

> The release was built from
> https://github.com/nextcloud/richdocuments/commit/2ed8a3421ab66e47d823

So rebasing the patch against the right revision fixes our integration
tests and the patch now also applies cleanly :-)

The conflicts we had in the patch were pretty minor and fixing them
didn't involve introducing new code/functionality.

[1]: https://github.com/nextcloud/richdocuments/issues/886

Signed-off-by: aszlig <aszlig@nix.build>