lool: Rebase patch for "richdocuments" app
The update to richdocuments version 3.6.0 introduced a few conflicts in
our patch but none of them are major, since the conflicts mostly
revolved around the fallback demo service.

This change only fixes the build for the richdocuments app but shouldn't
change anything in functionality.

I actually forgot to do this in da8610a50307961318bd36ff3980db23869d656,
alongside updating the app.

Signed-off-by: aszlig <aszlig@nix.build>