Fix SHA256 hash for "spreed" app
The tarball for spreed version 8.0.9 got re-uploaded after we updated it
in commit ac2a85157f70de82b9db050d092887384802212c.

To find out the reason for the re-upload, I diffed the contents of
previous tarball with the new one and the following files were added:

  js/admin
  js/admin/allowed-groups.js
  js/admin/allowed-groups.js.map
  js/admin/commands.js
  js/admin/commands.js.map
  js/admin/general-settings.js
  js/admin/general-settings.js.map
  js/admin/signaling-server.js
  js/admin/signaling-server.js.map
  js/admin/stun-server.js
  js/admin/stun-server.js.map
  js/admin/turn-server.js
  js/admin/turn-server.js.map
  js/collections.js
  js/collections.js.map
  js/flow.js
  js/flow.js.map
  js/talk-files-sidebar-loader.js
  js/talk-files-sidebar-loader.js.map
  js/talk-files-sidebar.js
  js/talk-files-sidebar.js.map
  js/talk-public-share-auth-sidebar.js
  js/talk-public-share-auth-sidebar.js.map
  js/talk-public-share-sidebar.js
  js/talk-public-share-sidebar.js.map
  js/talk.js
  js/talk.js.map

So it looks like the maintainer somehow forgot to either generate those
files or include them in the tarball.

Of course, when updating the SHA256 I also verified whether the
signature is correct via our updater.

Signed-off-by: aszlig <aszlig@nix.build>