tests/talk: Don't save browser log file
The log file doesn't seem to to exist and thus the test fails with an
error like this:

  sync: error opening '/tmp/xchg/browser.log': No such file or directory

Of course in the long term it's preferable to have the browser log,
beause it helps a lot when it comes to debugging. However if we need to
debug the test anytime soon, we can still figure out why the browser log
isn't written and subsequently fix it.

Signed-off-by: aszlig <aszlig@nix.build>