talk: Switch to using external signaling service
The external signaling service that so far has provided the so-called
"High Performance Backend" in Nextcloud has been open-sourced[1] a few
days ago.

This backend also has support for the Janus WebRTC server[2], which
ultimately will be our goal since it provides an SFU for more than just
a handful of participants. Right now however, this is not yet the case.

Nevertheless, since the external signaling backend uses Websockets
instead of long-polling, we should also have a much more responsive chat
experience already.

As usual, this comes with a small set of patches, which make sure that
we use Unix domain sockets everywhere and avoid the use of secrets
whenever possible and use peer authentication and file-based permissions
instead.

There are only two secrets (NEXTCLOUD_SIGNALING_HASHKEY and
NEXTCLOUD_SIGNALING_BLOCKKEY), which are used for signing the cookie of
the signaling backend. We probably could randomly generate those secrets
on every startup though, but right now it's persistent.

[1]: https://nextcloud.com/blog/open-sourcing-talk-back-end-rc-of-talk-9-brings-lots-of-improvements/
[2]: https://github.com/meetecho/janus-gateway

Signed-off-by: aszlig <aszlig@nix.build>