coturn: Rename/restructure into talk/ directory
Since the coturn/ directory no longer only contains STUN/TURN-related
content, using "talk" for "Nextcloud Talk" probably is a better name to
contain all the Nextcloud Talk related things.

With this small restructure, the coturn directory still exists within
talk/ but the external signaling along with its patches is now in a
dedicated directory, so it is more apparent which patches belong to the
corresponding service modules.

Signed-off-by: aszlig <aszlig@nix.build>