Add Nextcloud major version 19
Here is a short overview of the highlights:

  * Password-less authentication
  * Talk 9 with built-in office document editing, a grid view & more
  * Improved performance
  * Deck integration in Calendar
  * Guest account groups

Upstream release announcement:

https://nextcloud.com/blog/nextcloud-hub-brings-productivity-to-home-office/

Since this release, using OCP\Http\Client on localhost is no longer
allowed[1], so when communicating with Nextcloud via Unix domain
sockets, eg. from OSRM or the signaling backend we'll run into failures.

So to cope with this, I've rebased the signaling patch for Talk to use
"dummy.invalid" as the domain instead of localhost. Similarily I also
did the same for the Maps patch, however I didn't rebase it but simply
modified the patch file.

Since the "circles" app is not yet supported on Nextcloud 19 (the
release is done on GitHub[2] but hasn't been pushed to the app
store[3]), I've added it to the list of force-enabled apps for our
upgrade test.

I also added a constraint on the "ocsms" app for our updater, since the
Nextcloud 19 compatible version hasn't been released[4] and since
version 1.11.0 doesn't specify a max-version constraint[5] we'd end up
downgrading the app.

[1]: https://github.com/nextcloud/server/commit/5e402f8aaeacf05f956c6a73d7300e7849bc4bae
[2]: https://github.com/nextcloud/circles/releases/tag/v0.19.0
[3]: https://apps.nextcloud.com/apps/circles/releases
[4]: https://github.com/nextcloud/ocsms/issues/319
[5]: https://github.com/nextcloud/ocsms/blob/v1.11.0/appinfo/info.xml#L12

Signed-off-by: aszlig <aszlig@nix.build>