tests/upgrade: Exclude "camerarawpreviews" app
The app states in its info.xml file that it needs the "perl" command to
be enabled. This in turn leads to a failure during "occ app:enable" in
the test.

In the long term it makes sense to have a way to handle these
dependencies on our side, but for the time being, let's exclude the
camerarawpreviews app for now.

Signed-off-by: aszlig <aszlig@nix.build>