tests/urls/e2e: Use Nextcloud Desktop user agent
Since end_to_end_encryption version 1.5.2, the user agent header of the
client is checked by the server and if it doesn't match the user agent
of either one of the mobile clients or the desktop clients, the request
is rejected with an HTTP 403 Forbidden.

The API however is still the same, so all we need to do to fix this is
forge the user agent accordingly, so I went with the one from Nextcloud
Desktop on GNU/Linux.

Signed-off-by: aszlig <aszlig@nix.build>