Add Nextcloud major version 20
Since there are quite a lot of changes to be listed, here are the
highlights from the upstream blog post[1]:

  The three biggest features we introduce are:

    * Our new dashboard provides a great starting point for the day with
      over a dozen widgets ranging from Twitter and Github to Moodle and
      Zammad already available
    * Search was unified, bringing search results of Nextcloud apps as
      well as external services like Gitlab, Jira and Discourse in one
      place
    * Talk introduced bridging to other platforms including MS Teams,
      Slack, IRC, Matrix and a dozen others

  There is more:

    * Notifications and Activities were brought together, making sure
      you won't miss anything important
    * We added a 'status' setting so you can communicate to other users
      what you are up to
    * Talk also brings dashboard and search integration, emoji picker,
      upload view, camera and microphone settings, mute and more
    * Calendar integrates in dashboard and search, introduced a list
      view and design improvements
    * Mail introduces threaded view, mailbox management and more
    * Deck integrates with dashboard and search, introduces Calendar
      integration, modal view for card editing and series of smaller
      improvements

  Some other app improvements we want to highlight include:

    * Flow adds push notification and webhooks so other web apps can
      easily integrate with Nextcloud
    * Text introduced direct linking to files in Nextcloud
    * Files lets you add a description to public link shares

New external apps in Nextcloud 20:

  collectives -> default: false

    Collectives is a Nextcloud App for activist and community projects
    to organize together. Come and gather in collectives to build shared
    knowledge.

  flow_notifications

    Enable users to configure notifications with customized conditions
    in their Flow configuration.

  integration_{discourse,github,gitlab,google,jira,mastodon,
               moodle,reddit,twitter,zammad}

    Various integration apps for external services.

  talk_matterbridge

    Brings the Matterbridge binary to your server to connect Nextcloud
    Talk with other chat services.

In addition to those, we now have three more shipped apps that are
enabled by default: dashboard, user_status and weather_status.

The maintenance:install occ subcommand now requires to use the
--database-pass option and not doing so results in an error like this:

  What is the password to access the database with user <nextcloud>?

  Unable to hide the response.

Since we exclusively use peer authentication, we don't need to supply a
password, so setting the value of the option to an empty string causes
the error to go away.

Additionally, we now get the following exception message during
installation:

  Could not boot files_trashbinCould not resolve trashManager! Class
  trashManager does not exist

The error is non-fatal and doesn't seem to affect the finaly system, but
it's ugly nonetheless. Fortunately, there is already an upstream
issue[2], so we can expect this to be addressed in the next release.

One of the apps we are patching is gpxpod, which also got an update
specifically for Nextcloud 20. The update comes with a completely
changed directory structure that is more in line with the developer
manual, which is a good thing.

However, we do need to take this into account, when patching in
executable paths. Fortunately the code itself is unchanged though.

I also disabled the tencentcloudcosconfig app, since we currently get a
hash mismatch when fetching the app for Nextcloud 19 and I currently
have zero use for integrating this Tencent thingy.

Another app I disabled was the gpgmailer app, which needs GnuPG in order
to work. In the future we might want to properly patch apps like this
one to explicitly use a specific dependency.

When running the Nextcloud Cron service, the "news" app also fails with
an exception about "Call to a member function execute() on int". For our
VM test, this currently is not yet fatal, but in the long term we should
change this. Nevertheless, the issue also has been reported[3] upstream.

Nextcloud Talk now also has a few different CSS selectors, which I fixed
in our test driver.

Our proxy patch[4] for Nextcloud Maps also needed to be rebased against
version 0.1.8, which is only released for Nextcloud 20 in the app store.

[1]: https://nextcloud.com/blog/nextcloud-hub-20-debuts-dashboard-unifies-search-and-notifications-integrates-with-other-technologies/
[2]: https://github.com/nextcloud/server/issues/22590
[3]: https://github.com/nextcloud/news/issues/819
[4]: https://github.com/nextcloud/maps/pull/167

Signed-off-by: aszlig <aszlig@nix.build>