tests/upgrade: Temporarily disable officeonline
This app currently causes our upgrade test to fail because it uses the
rd_wopi_token_idx index, which is also used by the richdocuments app.

The issue has been fixed upstream[1] already, but no new version has
been released yet, so I added an assertion on version 1.0.0 so that once
we got the new upstream version, we can re-enable the app in our upgrade
test.

[1]: https://github.com/nextcloud/officeonline/pull/6

Signed-off-by: aszlig <aszlig@nix.build>