osrm: Add throttling for GraphHopper and Mapbox
This essentially adds the latest commit of my pull request[1] upstream
and throttles requests to GraphHopper and Mapbox to avoid hitting API
limits on these services.

Right now, the limit is 20 requests in 60 seconds, which is choosen
somewhat arbitrarily at the moment and the precise reason why I'm
including it here as well is to check whether users of my Nextcloud
instances are hitting these limits in production.

[1]: https://github.com/nextcloud/maps/pull/167

Signed-off-by: aszlig <aszlig@nix.build>