Implement basic Nix Flake
Right now, this is more like a patchwork implementation and it
essentially passes the pinned nixpkgs versions via the flake while also
allowing it to be used without going through the flake.

The main focus here is to have feature parity with what we had so far,
with a few exceptions.

For example, we have used Firefox 70 for testing, which apparently
worked fine in single process mode. However newer versions of Firefox
appear to block as soon as we want to start a WebRTC connection.

I didn't dig through all the changes, but my guess is that Firefox is
now handling this in a separate process, so switching to multi-process
mode fixes the test for Nextcloud Talk.

Another change is that I removed most of the legacy stuff in manual.nix
and essentially switched everything to the nixosOptionsDoc function.

In the long-term our goal is to reorganize the structure of the project
so that NixOS modules are separate from the actual packages.

Signed-off-by: aszlig <aszlig@nix.build>