flake.nix: Expose PostgreSQL module
So far the way one's supposed to use our drop-in PostgreSQL replacement
module is to directly specify its path. However, with flakes it's a
better idea to make the module available via the canonical nixosModules
attribute, since it's what users of this Flake would expect.

It's also shown in "nix flake show", so it also helps in
discoverability.

Signed-off-by: aszlig <aszlig@nix.build>