Add a default.nix for Flake compatibility
This is mostly for systems that do not yet have support for Nix Flakes
and I'd rather like to avoid this, but at the same time I try to run my
production servers with the latest NixOS stable version.

So until Nix Flakes are a stable Nix feature, let's use the
compatibility wrapper.

Signed-off-by: aszlig <aszlig@nix.build>