Temporarily add socket-timeout to uWSGI config
The default value for the socket timeout is 4 seconds, which is rather
low for large file uploads.

We really need to add a specific test for this, since setting the
socket-timeout to such a large value would probably make us susceptible
to DoS.

However, since the Nextcloud instances I'm running right now are mostly
internal, I've set it to 100000 for now, just to make sure file uploads
are working.

Signed-off-by: aszlig <aszlig@nix.build>