tests/upgrade: Temporarily excempt user_oidc app
Since the last round of Nextcloud major version upgrades in
dcd16d1554fb1c39d8502c95373d12214bf6a812, the user_oidc app now fails to
install with the following error:

  Column name "oc_user_oidc"."display_name" is NotNull, but has empty
  string or null as default.

This very message is the result of a new exception introduced[1] in
Nextcloud 19.0.4.2.

I also submitted an upstream issue[2] for the user_oidc app and added an
assertion on the version, so that we can re-enable the app for testing.

[1]: https://github.com/nextcloud/server/pull/22652
[2]: https://github.com/nextcloud/user_oidc/issues/90

Signed-off-by: aszlig <aszlig@nix.build>