tests: Temporarily remove "social" app from update
Unfortunately, the app currently fails to updated when going from
Nextcloud 19 to Nextcloud 20 with the following error:

  ERROR:  cannot drop index oc_social_3_cache_actor_pkey because
          constraint oc_social_3_cache_actor_pkey on table
          oc_social_3_cache_actor requires it
  HINT:  You can drop constraint oc_social_3_cache_actor_pkey on table
         oc_social_3_cache_actor instead.
  STATEMENT:  DROP INDEX oc_social_3_cache_actor_pkey

This is already reported upstream[1], so I added an assertion to make
sure we re-enable the app once a new version has been released.

Given that there are already a bunch of apps that are broken right now
in their versions from the Nextcloud app store, I think it might be a
good idea to soon provide warnings once people are trying to build a
NixOS config with a set of apps known to cause issues.

[1]: https://github.com/nextcloud/social/issues/1034

Signed-off-by: aszlig <aszlig@nix.build>