tests/upgrade: Switch to using nesting.clone
While nesting.clone got renamed to "specialisation", which is a better
name and also better interface (since it uses an attribute set instead
of a list), we can't yet use it since we're currently pinning NixOS
20.03.

Of course, we could run the tests using NixOS 20.09, but since
everything else is NixOS 20.03 the test would be less useful since it
won't detect breakages in NixOS 20.03.

Nevertheless, using nesting.clone already is a much better interface for
handling machine generations since we no longer need to rely on
internals of NixOS VM tests, nor do we need to add dummy machines.

The implementation can still be improved to provide a more sequential
datatype and allowing multiple major version upgrades, but I'll defer
that cleanup to after we switched to "specialisation" since it doesn't
make sense to build on an API that's going to be obsolete very soon.

I also removed "ldap_write_support" from the apps to be enabled, since
it only really works if there is an LDAP provider. If that's not the
case, the app doesn't even enable and will later result in a mismatch
when checking enabled apps against the one we expected to be enabled.

Signed-off-by: aszlig <aszlig@nix.build>