flake.lock: Override revision for NixOS 20.03
While switching our tests to the new Python-based NixOS VM test I
persistently ran into deadlocks of the log queue.

Since the fix[1] for this is is not yet part of the channel tarball, I
decided to manually override this since the latest evaluation is using
just the parent commit of that fix and I expect this to eventually
succeed anyway.

[1]: https://github.com/NixOS/nixpkgs/commit/b7bcf5e2477a9f10179d9c2ee89

Signed-off-by: aszlig <aszlig@nix.build>