updater: Don't create "result" output links
During the update process, we need to know the output store path of the
corresponding Nextcloud servers, so we use "nix-build" for that
accordingly.

Unfortunately, I forgot to pass --no-out-link, which means that every
time we run the updater we get a symlink called "result" in the current
directory.

We now not only pass --no-out-link but also set the "builders" option to
an empty string to prevent building remote.

Signed-off-by: aszlig <aszlig@nix.build>