Remove support for external theming
The Breeze Dark theme was the only one that was using the "theme" config
key and I specifically implemented a generic NixOS module option just
for that.

However, since Breeze Dark is shipped as a regular[1] Nextcloud app
since quite a while, it's really no longer necessary and we also didn't
include it as an external theme anymore since Nextcloud version 18.

With Nextcloud 17 gone, the "themes" attribute in our upstream.json
files became empty for all major versions, so we can safely get rid of
the attribute and all the code to support it.

[1]: https://github.com/mwalbeck/nextcloud-breeze-dark/commit/a17322cdf8d3253c30ee496ea09a7c9fed3c548d

Signed-off-by: aszlig <aszlig@nix.build>